requirements.txt
ag-ui-protocol==0.1.13
aiofile==3.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
argcomplete==3.6.3
attrs==25.4.0
Authlib==1.6.8
backports.tarfile==1.2.0
beartype==0.22.9
boto3==1.42.59
botocore==1.42.59
cachetools==7.0.1
caio==0.9.25
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cohere==5.20.7
cryptography==46.0.5
cyclopts==4.6.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
eval_type_backport==0.3.1
exceptiongroup==1.3.1
executing==2.2.1
fastavro==1.12.1
fastmcp==3.0.2
filelock==3.24.3
frozenlist==1.8.0
fsspec==2026.2.0
genai-prices==0.0.55
google-auth==2.48.0
google-genai==1.65.0
googleapis-common-protos==1.72.0
griffelib==2.0.0
groq==1.0.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.5.0
idna==3.11
importlib_metadata==8.7.1
invoke==2.2.1
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jeepney==0.9.0
jiter==0.13.0
jmespath==1.1.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
logfire==4.25.0
logfire-api==4.25.0
markdown-it-py==4.0.0
mcp==1.26.0
mdurl==0.1.2
mistralai==1.12.4
more-itertools==10.8.0
multidict==6.7.1
nexus-rpc==1.2.0
openai==2.24.0
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-httpx==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
packaging==25.0
pathable==0.5.0
platformdirs==4.9.2
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
py-key-value-aio==0.4.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-ai==1.63.0
pydantic-ai-slim==1.63.0
pydantic-evals==1.63.0
pydantic-graph==1.63.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
SecretStorage==3.5.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.3.2
starlette==0.52.1
temporalio==1.20.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
types-protobuf==6.32.1.20260221
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
wrapt==1.17.3
xai-sdk==1.7.0
yarl==1.22.0
zipp==3.23.0
